pyproject.toml
[tool.poetry]
name = "enunlg"
version = "0.2.0"
description = "Edinburgh Napier University package for Natural Language Generation"
authors = ["David M. Howcroft <dave.howcroft@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
regex = "^2021.4.4"
numpy = "^1.20.3"
bidict = "^0.21.2"
python-box = "^5.3.0"
jupyter = "^1.0.0"
ipython = "^7.25.0"
ipymd = "^0.1.3"
nbstripout = "^0.5.0"
# Uncomment the torch line relevant to your environment!
# Live dangerously and let poetry try to find something; this didn't work for me
#torch = "^1.9.0+cpu"
# Torch 1.9.1 on CPU-only for Python 3.9 on 64-bit Linux
#torch = {url="https://download.pytorch.org/whl/cpu/torch-1.9.1%2Bcpu-cp39-cp39-linux_x86_64.whl"}
# Torch 1.9.1 on CPU-only for Python 3.9 on 64-bit Windows
#torch = {url="https://download.pytorch.org/whl/cpu/torch-1.9.1%2Bcpu-cp39-cp39-win_amd64.whl"}
# Torch 1.9.1 on CPU-only for Python 3.9 on 64-bit Intel CPU MacOSX 10
#torch = {url="https://download.pytorch.org/whl/cpu/torch-1.9.1-cp39-none-macosx_10_9_x86_64.whl"}
# Torch 1.9.1 on CPU-only for Python 3.9 on 64-bit ARM MacOSX 11
#torch = {url="https://download.pytorch.org/whl/cpu/torch-1.9.1-cp39-none-macosx_11_0_arm64.whl"}
# Torch 1.9.1 on CPU-only for Python 3.9 on 64-bit Linux with CUDA 11
torch = {url="https://download.pytorch.org/whl/cu111/torch-1.9.0%2Bcu111-cp39-cp39-linux_x86_64.whl"}

hydra-core = "^1.1.1"
prettytable = "^3.4.1"
future = "^0.18.2"
sacrebleu = "^2.3.1"
lxml = "^4.9.2"
xsdata = {extras = ["cli", "lxml"], version = "^24.3.1"}
tensorboard = "^2.15.1"
setuptools = "59.5.0"
stanza = "^1.7.0"
submitit = "^1.5.1"
hydra-submitit-launcher = "^1.2.0"


[tool.poetry.group.dev.dependencies]
mypy = "^1.6.1"
pytest = "^5.2"
scalene = "^1.5.13"
types-regex = "^2023.10.3.0"
types-lxml = "^2023.10.21"
stanza = "^1.7.0"
ruff = "^0.3.4"


[tool.ruff]
exclude = [
    ".git",
]

line-length = 120
indent-width = 4

# Assume Python 3.9
target-version = "py39"

[tool.ruff.lint]
# Enable Pyflakes (`F`), the pycodestyle (`E`), and flake8-pathlib (`PTH`)  codes by default.
select = [
    "A",  # Shadowing builtins
    "B",  # Check for 'bugbears'
    "C",  # Use comprehensions comprehensively
#    "COM", # Check for weirdness with commas
#    "D",  # Check for docstrings & their styles
    "E",
    "EM",  # Check our error messages provide more useful tracebacks
    "F",
#    "FBT",  # Checks for the 'boolean trap' where a bool in a function call changes behavior
    "FURB",  # Tips for modernizing Python
    "ICN",  # Try to stick to common conventions when using aliases with imports, etc
#    "INP", # Implicit Namespace Package -- add __init__.py
#    "N",  # PEP 8 naming conventions
    "PERF",  # Performance tips linter (in preview)
    "PIE",  # Odds and ends
#    "PTH",
    "R",  # PyLint suggestions for refactoring
    "RET",  # Use return statements sensibly
#    "RUF",  # ruff-specific checks -- currently mostly flags confusable unicode chars which are as intended
    "SIM",  # Tips for simplifying code
]
ignore = ["E501"]

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
